# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Evently API",
    "description": "API for Evently",
    "version": "1.0.0"
  },
  "host": [
    "theodev.myftp.org:89"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/friends-request/": {
      "post": {
        "tags": [
          "Friends"
        ],
        "description": "Send a friend request to a user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Username of the friends",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "searchField": {
                  "type": "string",
                  "example": "azeraze#12345"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friend request sent successfully.",
            "schema": {
              "$ref": "#/definitions/FriendsRequest"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "404"
                },
                "msg": {
                  "type": "string",
                  "example": "User not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Friend request already exists or they are already friends.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "422"
                },
                "msg": {
                  "type": "string",
                  "example": "Already friends or request exists."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/friends-request/{username}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "description": "Remove a friend.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Username of the friends",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "azeraze#12345"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friend removed successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "200"
                },
                "msg": {
                  "type": "string",
                  "example": "Bien supprimé"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "404"
                },
                "msg": {
                  "type": "string",
                  "example": "User not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Users are not friends.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "422"
                },
                "msg": {
                  "type": "string",
                  "example": "They are not friends"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/friends-request": {
      "get": {
        "tags": [
          "Friends"
        ],
        "description": "Get friend requests for the logged-in user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of friend requests received by the logged-in user.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FriendsRequest"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/friends-request/{requestId}": {
      "put": {
        "tags": [
          "Friends"
        ],
        "description": "Accept or decline a friend request.",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the friend request to accept or decline",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Friend request acceptance data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "accept": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friend request accepted successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "isAccepted": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "boolean"
                    }
                  }
                },
                "user": {
                  "$ref": "#/definitions/User"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Friend request not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "404"
                },
                "msg": {
                  "type": "string",
                  "example": "Friend request not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/friends-request/sent": {
      "get": {
        "tags": [
          "Friends"
        ],
        "description": "Get friend requests sent by the logged-in user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of friend requests sent by the logged-in user.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FriendsRequest"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/friends-request/sent/{requestId}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "description": "Delete a sent friend request.",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the friend request to delete",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend request deleted successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "200"
                },
                "msg": {
                  "type": "string",
                  "example": "Friend request deleted successfully."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Friend request not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "404"
                },
                "msg": {
                  "type": "string",
                  "example": "Friend request not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/friends": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get user friends.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User friends retrieved successfully.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "name": {
          "type": "string",
          "example": "foobar"
        },
        "icon": {
          "type": "string",
          "example": "foobar"
        },
        "color": {
          "type": "string",
          "example": "foobar"
        },
        "default": {
          "type": "boolean",
          "example": true
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "Event": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "name": {
          "type": "string",
          "example": "foobar"
        },
        "description": {
          "type": "string",
          "example": "foobar"
        },
        "location": {
          "type": "string",
          "example": "foobar"
        },
        "childCreated": {
          "type": "boolean",
          "example": true
        },
        "recurrencePattern": {
          "type": "string",
          "example": "monthly"
        },
        "interval": {
          "type": "number",
          "example": 666
        },
        "maxOccurence": {
          "type": "number",
          "example": 666
        },
        "targetDate": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "inviteToken": {
          "type": "string",
          "example": "foobar"
        },
        "categoryID": {
          "type": "number",
          "example": 666
        },
        "userID": {
          "type": "number",
          "example": 666
        },
        "folderID": {
          "type": "number",
          "example": 666
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "deletedAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "joinedUser": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JoinedEventEntity"
          }
        }
      }
    },
...
    "FriendsRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "asker": {
          "$ref": "#/definitions/User"
        },
        "receiver": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "JoinedEventEntity": {
      "type": "object",
      "properties": {
        "leaveDate": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "invitationStatus": {
          "type": "string",
          "example": "invited"
        },
        "event": {
          "$ref": "#/definitions/Event"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "username": {
          "type": "string",
          "example": "foobar"
        },
        "firstName": {
          "type": "string",
          "example": "foobar"
        },
        "lastName": {
          "type": "string",
          "example": "foobar"
        },
        "email": {
          "type": "string",
          "example": "foobar"
        },
        "sub": {
          "type": "string",
          "example": "foobar"
        },
        "password": {
          "type": "string",
          "example": "foobar"
        },
        "provider": {
          "type": "string",
          "example": "foobar"
        },
        "isDeleted": {
          "type": "boolean",
          "example": true
        },
        "isCompleted": {
          "type": "boolean",
          "example": true
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "deletedAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Event"
          }
        },
        "joinedEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JoinedEventEntity"
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}